requirements.txt
python-dotenv
google-generativeai
Pillow
PyPDF2
streamlit
pytesseract